Look the max value of just the two values under the last one
Instead of looking max value in the whole row, we start with the row, drop as many columns as the column of the last value, take just 2 values, and compute the max of them.

Then, find the index of that max value (it can be just 0 or 1).

Next iteration, the column index is the last one plus the index of the current max value.